+++ analyzer/tool/analysis_driver/fuzz1.dart
@@ -1,13 +1,15 @@
 import 'package:analyzer/file_system/memory_file_system.dart';
 import 'package:analyzer/src/dart/analysis/analysis_context_collection.dart';
-import 'package:analyzer/src/test_utilities/mock_sdk.dart';
 
 main(List<String> args) async {
   final input = args[0];
 
   var resourceProvider = MemoryResourceProvider();
 
-  MockSdk(resourceProvider: resourceProvider);
+  // createMockSdk(
+  //   resourceProvider: resourceProvider,
+  //   root: resourceProvider.newFolder('/sdk'),
+  // );
 
   var path = '/home/test/lib/test.dart';
   resourceProvider.newFile(path, input);
@@ -15,7 +17,7 @@ main(List<String> args) async {
   var collection = AnalysisContextCollectionImpl(
     resourceProvider: resourceProvider,
     includedPaths: [path],
-    sdkPath: sdkRoot,
+    sdkPath: '/sdk',
   );
   var session = collection.contextFor(path).currentSession;
   await session.getResolvedLibrary(path);