+++ analyzer/tool/analysis_driver/fuzz2.dart
@@ -1,14 +1,16 @@
 import 'package:analyzer/file_system/memory_file_system.dart';
 import 'package:analyzer/src/dart/analysis/analysis_context_collection.dart';
 import 'package:analyzer/src/dart/analysis/file_byte_store.dart';
-import 'package:analyzer/src/test_utilities/mock_sdk.dart';
 
 main(List<String> args) async {
   final input = args[0];
 
   var resourceProvider = MemoryResourceProvider();
 
-  MockSdk(resourceProvider: resourceProvider);
+  // createMockSdk(
+  //   resourceProvider: resourceProvider,
+  //   root: resourceProvider.newFolder('/sdk'),
+  // );
 
   var path = '/home/test/lib/test.dart';
   resourceProvider.newFile(path, input);
@@ -16,7 +18,7 @@ main(List<String> args) async {
   var collection = AnalysisContextCollectionImpl(
     resourceProvider: resourceProvider,
     includedPaths: [path],
-    sdkPath: sdkRoot,
+    sdkPath: '/sdk',
     byteStore: FileByteStore(
       '/usr/local/google/home/scheglov/dart/fuzz2/bin/byte_store',
     ),